requirements.txt
black==19.10b0
celery==4.4.7
dj-database-url==0.5.0
django-compressor==2.4
django-debug-toolbar==2.2
django-dynamic-models-readonly==0.1.1
django-environ==0.4.5
django-extensions==2.2.9
django-redis==4.12.1
django-sass-processor==0.8
django-staff-sso-client==2.1.0
django-webpack-loader==0.7.0
django==3.0.7
factory-boy==2.12.0
flake8-import-order==0.18.1
flake8==3.7.9
geoalchemy2>=0.8.4
libsass==0.20.0
paste==3.4.3
psycopg2==2.8.4
pytest-cov==2.8.1
pytest-django==3.9.0
pytest-mock==3.1.0
pytest==5.4.1
python-dotenv==0.12.0
redis==3.5.3
sentry-sdk==0.16.1
sqlalchemy>=1.3.15
waitress==1.4.3
Werkzeug==1.0.1
whitenoise==5.0.1
XlsxWriter==1.2.8